rename files to dart convention
