+++ packages/flutterfire_cli/lib/src/command_runner.dart
@@ -17,10 +17,10 @@
 
 import 'package:args/command_runner.dart';
 
-import 'commands/bundle-service-file.dart';
+import 'commands/bundle_service_file.dart';
 import 'commands/config.dart';
 import 'commands/update.dart';
-import 'commands/upload-symbols.dart';
+import 'commands/upload_symbols.dart';
 import 'common/utils.dart';
 import 'flutter_app.dart';
