+++ packages/flutterfire_cli/lib/src/commands/bundle_service_file.dart
