+++ packages/flutterfire_cli/lib/src/commands/upload_symbols.dart
